Add files via upload
This plug-in convolves the two channels of a selection from a stereo track.  A typical application is to convolve a monaural audio selection on one channel with an impulse function on the other which represents a desired reverberant environment.  Note that the  Nyquist (convolve( function is not a fast one and subject to memory limits.  Experimentally, it was able to convolve a 20 second selection at 48KHz sample rate with a 5 second impulse in about 30 minutes on a Dell Laptop with an i5-5200 CPU.  YRMV.  

Attempts to convolve a larger selection will throw an immediate error which will appear in the debug window.  Attempts to convolve a monaural track, convolve between tracks, or convolve channels of unmatched sampling rates will also error.  

A successful convolution is placed identically in both channels of the source track.   

Depending on the nature of what is being convolved and the scale factors applied, the convolve function can easily overflow.  Before you commit to 30 minutes of convolving, you may wish to set both scale factors to 50% and convolve a short, but representative selection of your raw material, then reduce the scale factors as may be necessary to avoid overflow and clipping.